requirements.txt
pyusb
libusb
slackclient
yaml
glob